ssss

+++ db/migrate/20230811131923_addtochangecolumn.rb
@@ -1,7 +1,9 @@
 class Addtochangecolumn < ActiveRecord::Migration[7.0]
   def change
     # First, remove the existing columns
-    remove_column :employees, :location_id
+    if column_exists?(:employees, :location_id)
+      remove_column :employees, :location_id
+    end
     remove_column :employees, :department_id
 
     # Then, add the references columns
